bin/cli/requirements.txt
pycryptodome
# Copyright 2020 Pants project contributors.
# Licensed under the Apache License, Version 2.0 (see LICENSE).

ansicolors==1.1.8
setuptools>=56.2.0,<57
types-setuptools>=56.2.0,<58
